% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignBased.R
\name{calc_RMSE}
\alias{calc_RMSE}
\title{Function to calculate root-mean-squared-error for a model}
\usage{
calc_RMSE(pred, obs)
}
\arguments{
\item{pred}{Predicted CPUE from a model}

\item{obs}{CPUE observations}
}
\description{
This function calculates the rmse for a model given the observed and predicted
observations. It was stolen from Brian Stock.
}
\examples{
calc_RMSE(glm.cpue$fitted.values,Juvenile_POP_Data$juvenile_POP_CPUE)
}
\keyword{evaluation}
\keyword{model}
