% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa.strata.area}
\alias{goa.strata.area}
\title{goa.strata.area}
\format{A data frame with 59 rows and 4 variables:
\describe{
  \item{STRATUM}{unique strata identifier from RACEBASE}
  \item{AREA_KM2}{area of stratum in km2}
  \item{PROP_TRAWL}{proportion of strata that is trawlable}
  \item{DEPTH_BIN}{depth identifier}
  ...
}}
\source{
{RACEBASE}
}
\usage{
goa.strata.area
}
\description{
Table of stata area for the Gulf of Alaska
}
\keyword{datasets}
