% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Juvenile_POP_Data}
\alias{Juvenile_POP_Data}
\title{Juvenile Pacific ocean perch CPUE 
This is a data set containing catch per unit of effort for juvenile (< 250 mm) Pacific 
ocean perch in the Gulf of Alaska bottom trawl survey from 1996-2009. The data originated
from RACEBASE and includes the midpoint of the latitude and longitude of the tow (lat & long)
the survey year, the thermocline depth,the bottom water temperature, the bottom water 
depth, the seafloor slope at the survey location, the combined catch of structure forming 
invertebrates (sponges and corals), the water temperature at the thermocline depth, the 
combined catch of shrimp species in the bottom trawl haul and finally catch per unit of 
effort for juvenile POP.}
\format{An object of class \code{data.frame} with 4475 rows and 12 columns.}
\source{
{RACEBASE}
}
\usage{
Juvenile_POP_Data
}
\description{
Juvenile Pacific ocean perch CPUE 
This is a data set containing catch per unit of effort for juvenile (< 250 mm) Pacific 
ocean perch in the Gulf of Alaska bottom trawl survey from 1996-2009. The data originated
from RACEBASE and includes the midpoint of the latitude and longitude of the tow (lat & long)
the survey year, the thermocline depth,the bottom water temperature, the bottom water 
depth, the seafloor slope at the survey location, the combined catch of structure forming 
invertebrates (sponges and corals), the water temperature at the thermocline depth, the 
combined catch of shrimp species in the bottom trawl haul and finally catch per unit of 
effort for juvenile POP.
}
\keyword{datasets}
