% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMGAMRF.R
\name{predict.gam.bindex}
\alias{predict.gam.bindex}
\title{A function to predict CPUE-index from a delta-lognormal GAM with bootstrapped errors}
\usage{
\method{predict}{gam.bindex}(pa.model, cpue.model, boot_reps = 500)
}
\arguments{
\item{pa.model}{binomial GAM predicting presence or absence}

\item{cpue.model}{log-normal GAM predicting abundance where catch is positive}

\item{boot_reps}{number of bootstrap replications (default = 500)}
}
\description{
This function uses a binomial GAM and a log-normal GAM to compute a delta-lognormal
index of abundance with associated 95% confidence intervals estimated by the boot-
strapping the data. A dummy variable for year must be included to get the annual abundance
index. The function back-transforms the index and the CI's.
}
\examples{
gam.pa.xvars<-c("inverts","slope","btemp","bdepth")
gam.pa.yvar<-ifelse(PA.data[species.name]>0,1,0)
gam.pa.form <- as.formula(paste("gam.pa.yvar ~", paste(gam.pa.xvars,collapse="+"),"+as.factor(year)",sep=""))
pa.gam <- gam(gam.pa.form, family = binomial, data = PA.data)

CPUE.data<-subset(PA.data,PA.data[species.name]>0)
gam.cpue.yvar<-unlist(log(CPUE.data[species.name]))
gam.cpue.xvars<-c("inverts","slope","btemp","bdepth")
gam.cpue.form <- as.formula(paste("gam.cpue.yvar ~", paste(gam.cpue.xvars,collapse="+"),"+as.factor(year)",sep=""))
cpue.gam <- gam(gam.cpue.form, family = gaussian, data = CPUE.data)

predict.gam.bindex(pa.gam,cpue.gam)

}
\keyword{abundance}
\keyword{bootstrap}
\keyword{delta-lognormal}
\keyword{gam,}
\keyword{index,}
\keyword{survey}
