% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMGAMRF.R
\name{predict.glm.bindex}
\alias{predict.glm.bindex}
\title{A function to predict CPUE-index from a delta-lognormal GLM with bootstrapped errors}
\usage{
\method{predict}{glm.bindex}(pa.model, cpue.model, boot_reps = 500)
}
\arguments{
\item{pa.model}{binomial GLM predicting presence or absence}

\item{cpue.model}{log-normal GLM predicting abundance where catch is positive}

\item{boot_reps}{number of bootstrap replications (default = 500)}
}
\description{
This function uses a binomial GLM and a log-normal GLM to compute a delta-lognormal
index of abundance with associated 95% confidence intervals estimated by the boot-
strapping the data. A dummy variable for year must be included to get the annual abundance
index. The function back-transforms the index and the CI's.
}
\examples{
glm.pa.xvars<-c("inverts","slope","btemp","bdepth")
glm.pa.yvar<-ifelse(PA.data[species.name]>0,1,0)
glm.pa.form <- as.formula(paste("glm.pa.yvar ~", paste(glm.pa.xvars,collapse="+"),"+as.factor(year)",sep=""))
pa.glm <- glm(glm.pa.form, family = binomial, data = PA.data)

CPUE.data<-subset(PA.data,PA.data[species.name]>0)
glm.cpue.yvar<-unlist(log(CPUE.data[species.name]))
glm.cpue.xvars<-c("inverts","slope","btemp","bdepth")
glm.cpue.form <- as.formula(paste("glm.cpue.yvar ~", paste(glm.cpue.xvars,collapse="+"),"+as.factor(year)",sep=""))
cpue.glm <- glm(glm.cpue.form, family = gaussian, data = CPUE.data)

predict.glm.bindex(pa.glm,cpue.glm)

}
\keyword{abundance}
\keyword{bootstrap}
\keyword{delta-lognormal}
\keyword{glm,}
\keyword{index,}
\keyword{survey}
