% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestIndex.R
\name{predict.rf.index}
\alias{predict.rf.index}
\title{A function to predict CPUE-index from a random forest model with bootstrapped errors}
\usage{
predict.rf.index(data, rf.yvar, rf.form, rf.model, boot_reps = 500)
}
\arguments{
\item{data}{data used to formulate the random forest model}

\item{rf.yvar}{y variables (CPUE) in the random forest model}

\item{rf.form}{formula for the random forest model}

\item{rf.model}{random forest model to bootstrap}

\item{boot_reps}{number of bootstrap replications (default = 500)}
}
\description{
This function uses a random forest model to produce and index of abundance with 
associated 95% confidence intervals estimated by the boot-strapping the data. 
A dummy variable for year must be included to get the annual abundance
index. The function back-transforms the index and the CI's.
}
\examples{
rf.cpue <- randomForest(rf.form,data=PA.data, mtry=3, ntree=1000, importance=TRUE, do.trace=250, keep.forest=TRUE) 
print(rf.cpue)	
rf.index<-predict.rf.index(PA.data,rf.yvar,rf.form,rf.cpue)

}
\keyword{abundance}
\keyword{bootstrap}
\keyword{forest,}
\keyword{index,}
\keyword{random}
\keyword{survey}
