% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMGAMRF.R
\name{rep.col}
\alias{rep.col}
\title{Function to repeat the columns of a matrix}
\usage{
\method{rep}{col}(x, n)
}
\arguments{
\item{x}{matrix or data frame}

\item{n}{number of times to repeat the columns of the matrix x}
}
\description{
This function repeats the columns of a matrix, x for n times
}
\examples{

}
\keyword{repeat}
\keyword{rows}
