% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMGAMRF.R
\name{rep.row}
\alias{rep.row}
\title{Function to repeat the rows of a matrix}
\usage{
\method{rep}{row}(x, n)
}
\arguments{
\item{x}{matrix or data frame}

\item{n}{number of times to repeat the rows of the matrix x}
}
\description{
This function repeats the rows of a matrix, x for n times
}
\examples{

}
\keyword{repeat}
\keyword{rows}
