% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignBased.R
\name{Stratified_CPUE}
\alias{Stratified_CPUE}
\title{Function to estimate Biomass or CPUE from stratified-random survey design}
\usage{
Stratified_CPUE(Catch, Year, Strata, Strata_area, Region = "GOA",
  Proportion_trawlable = 1, method = "Biomass")
}
\arguments{
\item{Catch}{Catch-per-unit-of-effort for each bottom trawl haul}

\item{Year}{Survey year}

\item{Strata}{Strata designation for each haul}

\item{Strata_area}{Area of the strata in km2 where each haul occurs}

\item{Region}{Either "GOA" (the default) or "AI"}

\item{Proportion_trawlable}{The proportion of the strata that can be sampled by the survey
trawl (default is 1)}

\item{method}{Either "Biomass" or "CPUE"}
}
\description{
This function estimates a design-based survey estimator for a stratified-random survey 
design. Either a population biomass or CPUE can be output. At the moment the function is
set up to incorporate the Gulf of Alaska or the Aleutian Islands bottom trawl survey.
It requires CPUE data, Strata designations and total area (in km^2). It can also be used 
to calculate an estimate for only the trawlable portion of the strata.
}
\examples{
design.index<-Stratified_CPUE(Design.data$juvenile_POP_CPUE,Design.data$year,Design.data$STRATUM,Design.data$AREA_KM2,"GOA",1, "CPUE")
design.index<-Stratified_CPUE(Design.data$juvenile_POP_CPUE,Design.data$year,Design.data$STRATUM,Design.data$AREA_KM2,"AI",Design.data$Trawlable, "Biomass")
}
\keyword{abundance}
\keyword{index}
\keyword{random}
\keyword{stratified}
\keyword{survey}
\keyword{survey,}
